Dcs 52 enter use of force details (#33)
* DCS-52: initial creation of nunjucks components in Details page. Incident.js updated to inject content into components

* DCS-52: validation messages

* DCS-52: correct misspelling of baton

* DCS-52: Cypress tests

Altered form input names to camel case and renamed some to be consistent with others.
Created new Cypress tests that make radio button and checkbox selections. The values of the selections will be verified in the database.

* DCS-52: renamed element names to camel case

* DCS-52:  radio buttons displayed inline rather than stacked. Custom nunjucks macros written as form-builder macros did not provide the required functionality

* DCS-52: change to cypress test

* DCS-52: refactor Cypress tests

* DCS-52: renamed macro and removed reference to nodemon in package.json

* DCS-52 Updated nunjucks macro

The macro previosuly needed the content to be passed in as args. These were difficult to write so extraxcted the args into a json object and passed that object in as the single arg

* DCS-52: Replaced inline Style attribute with CSS class

Previously used inline CSS in details.html to clear floats. Replaced this by creating a new class called clear-both in local.sass

* DCS-52: Input data now in JSON form and Macros reformatted to use loops rather than fixed elements so that variable length inputs can be used

* DCS 52 ameneded typo

* DCS 52: removed linter errors

* DCS 52: method next() changed to save() as per master and cypress tests updated