DCS-52 bugfix to remove H1's from legends
A HTML validation check revealed we can not place H1's inside legends.
The Govuk components do however allow them but only if the H1 represents the page header.
This would happen if there is one question per page and the header and question are the same.
In the UoF pages we have multiple questions on each page hence I have removed the H1s leaving the legend to represent the question for the relevant input component.