DCS 52: Users selections repopulated when moving between pages
http://localhost:3000/form/incident/details/-35

The Details page was saving user selections to DB but if user navigated back to the same page,
their previous selections were not being presented.
There is a global object called data that holds all the selection info so have used that inside incidentMacro.njk and
tested it against the value property in details.html to display the previous selections.