DCS 52: Users selections repopulated when moving between pages (#44)
* DCS 52: Users selections repopulated when moving between pages

http://localhost:3000/form/incident/details/-35

The Details page was saving user selections to DB but if user navigated back to the same page,
their previous selections were not being presented.
There is a global object called data that holds all the selection info so have used that inside incidentMacro.njk and
tested it against the value property in details.html to display the previous selections.

* DCS-52: Previous user selections now displayed in UI

http://localhost:3000/form/incident/details/-35

The logic {% if option.value in question.followUpQuestion.value %} checked  {% endif %}> in incidentMacros.njk (around line 86) checks the restraintPositions array held in DB in order to assign a 'checked' attribute to inputs. This is needed to show a user what selections they had made on the details.html page once they had passed through it.
However on the first visit to the details page, the user wouldn't yet have made a selection hence the logic was trying to test for a value in an undefined array object. This caused the application to fail.
To get around this, have amended details.html to assign either the  value retrieved from the DB or an empty array, to the followUpQuestion.value for question 6.