DCS-61: created a sanitiser function
Previous sanitiser function was difficult to read and was being defined mutiple times.
Essentially the function takes an array of objects and removes any elements that have an empty name value.
Created a new function called inputsExcludingEmptyStrings() that is easier to understand.
It is defined once and called multiple times.