DCS-74: Date and Time of incident displayed
http://localhost:3000/check-answers/-1

The Check Answers page now displays 2 new rows for date and time.
The Cypress tests use regex to verify the format of the date and time correspond to DD/MM/YYY and HH:mm. The tests don't check the actual date and time because the date and time are generated when the tests are run hence they can't be hard-coded in advance.