DCS-74: Date and time should have correct format
The Cypress tests previously generated a regex of the date and time for the test. Unfortunately rather than generating DD/MM/YYY and HH:mm, it actually generated the current date and time, eg 05/08/2019 and 13:47.
This meant that although the tests were passing they were doing so on an incorrect basis.
So have revised the regex and used the invoke('trim'), which is just like Jquery's trim () function to remove any hidden characters before and after the date and time