DCS-38 Adding ability to retrieve emails for usernames
This is not currently hooked up to anything and also does not take into account users
which exist in nomis but not new-nomis.
(For the alpha we can ensure all users will exist in new-nomis but will ticket up this limitation)

It groups usernames into those that have verified email address, those that don't and those that
don't exist. We'll use these groups and the success flag to determine whether we can persist the
data and how to display validation messages if we can't