DCS-86: Validation summary messages are displayed in Evidence page (#87)
http://localhost:3000/form/incident/evidence/-1

The user will now see the validation summary box with the appropriate links if they do not make a response to the questions in the Evidence page.
Have only included summary messages at this point. Will do the code for displaying individual validation messages above each question as a separate ticket as it is a more involved piece of work.

Cypress integration tests updated and unused variables removed following linting.