DCS-75 Adding validation to statement page
Always persistng statement on submit - even if valdiation fails.
This means that a user can comeback to their work later, even if not complete
and we can avoid issues with trying to store large amounts of content in cookie sessions.

Not carrying out validation when the user click save and return to incident list.