DCS-100 Adding a report status checker (#136)
This checks that each section of a report is complete and reports back either:
COMPLETE, INCOMPLETE or NOT_STARTED

This will then be used to determine whether to show the 'check your answers' page
and also to ensure that we can do a checkState when submiting a report.

Also making involved staff optional as this now refers to staff other than the
reporting officer. (If the use of force only involved the reporter than no other staff
would be entered)

Also adding a bit of an extra check to ensure that users cannot add involved staff
details through nefarious means.