DCS-105: user can exit after creating report back to incidents page
http://localhost:3000/74/report-sent

New button (link) allowing user to exit from the journey and go to DPS homegage rather than go on to complete the statement. The fallback to going to DPS home is the incidents page.
Config.js and helm updated so that appropriate url for DPS home is provided dependant of the environment user is accessing.