DCS-118: User is now able to view the current statements before moving to an Add Comments page
http://localhost:3000/reportId/add-comment-to-statement

The previous journey was for user to view and amend the Statement on the same page.
The button to access the Statement page was on the Incidents page
The journey has been changed so that the the Incidents page only has a view button. This takes user to the Statement page which displays the initial statement plus any additional comments.
From here the user can either click Continue to go back to the Incidents page or click Add comment to go to the new page.
The new page displays the same information as the Statement page but also has a text box to accept another comment.
The user then has the option to either Save a new comment (and go back to the Incidents page) or Cancel and return to the previous page.