DCS-93 Processing reminders
The logic to work out which type of reminder to send based on whether the recipients
was the reporter and whether they are overdue or not.

I've also updated the email templates and added an environmental variable for the link
which appears in the emails.

After being processed, items should become ineligible for subsequent polls by either
having their next reminder date push to the next day, or if they statement is overdue
then the date will be blanked out to show that no further reminders will occur.

I've temporarily reduced the amount of items that are processed in one go, as don't potentially
want 100s of emails being sent whilst testing.