DCS-93 seeding initial next reminder date
This seeds the initial next reminder date when the report is submitted.
The initial reminder date is the submitted date plus one.

Also fixing an issue where we weren't returning incident date as part
of the reminder item which meant the email template was defaulting
incident date and time to 'now'