DCS-114: Staff Involved usenames capitalised (#161)
localhost:3000/report/-1/incident-details

Have capitalised the usernames during the input sanitisation process. This was required because previously, the user could  type in a same username multiple times simply by using differing character case. This resulted in the insame the person being displayed multiple times in the Check Answers page. Capitalising the username has meant the equivalence can be identified and thus either display a validation message or display the IN PROGRESS marker in the tasklist.'

Cypress tests updated because as the  uppercased username will appear application wide.