Tidying up tests
Adding the ability to seed reports - this means we can write tests
that don't have to go through the entire application to work on a submitted
report.

There is one test, 'form submit/Submitting a form' which checks that the payload
we used to seed the db with, matches the generated payload from actually going
through app.