dcs-116: initial unit tests to validate  involvedStaff, witnesses and staffNeedingMedicalAttention
The fields across the application that accept either a username or a person's name were previously not checking for unexpected characters.
Have now introduced regex so that usernames only contain numbers, letters and an underscore. And person names only contain letters, spaces, hyphens, dots and apostrophes.  Min and max lengths for both are 2 and 50 chars.