DCS-115: Browser page tab displays the same message as page title
The browser tab was a generic GOV.UK statement previously. Have now matched it to the title of the page the tab relates to.
Have applied a generic tab message of 'GOV.UK - Use of force' as a default should there ever be a page where the title isn't set using Nunjucks