DCS-160: fixed blue section in page header (#218)
* DCS-160: fixed blue section in page header
The blue section was previously slightly taller than the surrounding black bar so have reduced its height. Have also chaned the layout when screen is Samsung S5 sized. The only 2 screen sizes needed to be considered for are Samsung S5 and desktop
Note have changed the blue colour to shade #015ea5 to be consistent with DPS. Could not find a sass code equivalent to $govuk-blue in the latest GDS toolkit hence have used the hex code

* DCS-160: hex value for colour changed to sass variable
Hex value for blue replaced with a sass variable. Sass variable defined in a new file called palette.sass. Have also redefined spacing around the navigation links so that when screen size is as for S5, the links can all fit across one line