DCS-223 Adding temporary endpoint to add missed staff
We have already received 5 requests to add involved staff members to a report as they
were missed when the report was raised.

We have a manual process to do this but its fiddly, easy to get wrong and very time consuming.
It also involves manually running db statements.

The idea is to add the ability for coordinators to be able to do this through the
application which will require input from design and additional time.

As a temporary fix, this change adds an endpoint that can be called by internal staff to add
these member easily.

We don't have any auth setup for api style endpoints so I've made this easy to trigger from
a browser as a logged in user.