DCS-320 Fix adding broken involved staff
It is is possible to add an invalid member of staff by passing a username in the invalid case.
The staff lookup performs a case insensitive look up but we end up persisting the user provided value.

When we retrieve statements for the current user, we will do a case sensitive lookup based on the current username.
It is likely that the two versions of the usernames don’t match up.

This changes it so that we always persist the service provided username