DCS-376 Allow user to alway complete reports
Currently a user is prevented from continuing to complete a report if the offender leaves the current prison.
This change will allow users to always see offender details if the report already exists.
(We need to prevent prison officers from seeing offender details if they do not exist in the current establishment)

This will also enable the posibility of a monitored process that will create a draft report for a prisoner that has
already left an establishment (potentially by a coordinator). The reporter will be able to continue the draft without
being given access to other caseloads.