DCS-387 Adding monitoring and tweaking job
Altering job to use the replace concurrency policy. This should kill an existing job if it lasts longer
than the next scheduled start time and start a new pod.

Also setting a deadline to ensure that the job ends if it lasts longer than 4 minutes

(The job runs every 5 minutes and processes up to 50 reminders which is generally very fast)

Also adding 2 prometheus checks:
1.) to check that the previous schedule time was less than 10 minutes
2.) to check that no recents jobs failed