DCS-394 Ensure retrieve correct incident location list
This change ensures that we only retrieve internal locations from the establishment where
the incident occured, rather than the establishment that the offender currently is.

The internal location list is only needed when the draft report is being filled out but
if creating a report for an offender who has recently left prison this distinction is
important.

This change also removes loading location list when unnecessary - we previously retrieved
the list every time we accessed the offender details.