DCS-237: conditional check should have tested for undefined
The pain inducing techniques question should only be absent in the check your answers page for old reports. They wouldn't have been asked the question so it will not be in the DB.
Any new reports still need to display the question irrespective of whether the user selected Yes or No when completing the incident pages.