DCS-236 'Was the prisoner compliant' question enhanced
http://localhost:3000/report/-19/relocation-and-injuries

The question has been enhanced so that if the user selects No, they are presented with 3 radios buttons each of which relate a different question.

If they had selected 'Yes' then there would be no supplementary radios.

The 'check your answers'  and the 'your-report' also
display the choices made to the supplementary question.