DCS-208 Adding heat map report
This shows locations by number of incidents at a specific hour
This is to allow coordinators to be able to determine which parts of a prison
have the most incidents of usages of force.

This will produce a CSV where location rows are sorted to show the most number
of incidents first and then using order of location name to resolve ties.