DCS-472 Fix issue adding unverified users
If an unverified user is added then the system was storing
the user provided version of the username.

This causes an issue where elite2 and auth use case insensitive search so
return results correctly but we then cannot load their statement as we do a case
sensitive search on statements by user id.

We'd rather store the user ids in the canonical form so upper casing here.