DCS-583: User was previously able to input 23:59 minutes in current d… (#388)
* DCS-583: User was previously able to input 23:59 minutes in current day without it failing future time validation.
This was because the timeValidation() function was working with minutes rather than seconds.
Function changed to use seconds.
The user could also previously enter negative numbers for hour and minutes. Validation added in minuteValidation() and hourValidation() to identify values less than zero.
dateSanitiser stoDateTime() updated to invalidate negative numbers

* DCS-583- test updated

* dcs-583: change to validation messages

* dcs-583: test to verify an error message is displayed if time input is 23:59 of current day

* DCS-583: validation to capture letters mixed with numbers for hours and minutes
The js parseInt() function will return a number if the first char is a number irrespective of the following chars. So regex used to
only run the parseInt() if the input is all numbers.

* DCS-583 regex for toInteger() changed

Co-authored-by: G <gurnank@outlook.com>