DCS-618: timeInput macro refactored (#392)
* DCS-618: timeInput macro refactored
The macro has been trimmed down to remove any unused elements. The incidentsDetails.html file uses this module so that has been updated to pass in only the changing content (value and error classes). All the other classes and attributes for the hour and minute has been hard coded into the macro.
Have added a new class to widen the date input field because when there is an error in the date, the red errors border covers some of the field content.

* DCS-618: refactor and file name change following PR comments

Co-authored-by: G <gurnank@outlook.com>