DCS-403 Removing missing user check
Previously we would prevent a user from submitting a report if they
had listed users in the report that didn't exist in nomis.

This was to handle an edge case where the user failed to resolve
a missing user after being prompted.

Going forward there will be no way to add a missing staff member so
this check will not be necessary. Implementing this change now
to ease moving to new flow