DCS-640 Tidying up application creation
Separating out a service index.ts which does the service wiring up from
the creation of the app.

The type for the services container can then be passed straight through to the routers
rather than threading through each service independently.

This had a knock on effect of forcing tests to be tidied up a bit