DCS-910: email will be sent when involved staff removed (#509)
There are two ways an involved staff can be removed from an incident report.
First is by coordinator using the 'delete' link in view-report page and the second
is the coordinator using the 'view removal request' link in visa-statements and following the process through.
Ultimately they both do the same thing. The action to send an email to the involved staff member starts in the involvedStaffService.removeInvolvedStaff() method.
A new template has been created in govuk Notify.