DCS-986 removed the feature toggle for disputed involvement (#525)
The feature toggle was applied to determine the content of emails sent to involved staff. The new content included a link the user could click in order to request they be removed from the UoF report. 
An env var called DISPUTE_INVOLVEMENT_ENABLED had been set to true for all envireonemnts that needed to use the new email templates. Because the templates are now needed by dev, pre and prod, there is no need to have a feature toggle anymore. So DISPUTE_INVOLVEMENT_ENABLED removed from helm and the associated disputeInvolvementEnabled removed from config.js