DCS-1001-bugfix for username (#532)
Application error if coordinator attempted to add an involved staff who was already in the report. Error only happened if the username was input in lower or mixed case rather than all uppercase.
Fix was to supply the uppercased username string when checking if user was already in the report.