Map 2319 uof redesign report pages to include tab (#784)
* MAP-2319 initial commit for reviewer view

* MAP-2319 more css for view-report and update unit test

* MAP-2319 redesign /view-report and /your-report to show a Report tab, data in sections, sortable table and conditional content.
The /check-your-answers page shares the same underlying reportDetailMacro.njk used by the report pages, however its design is not to be changed.
Hence the use of a renderReportView variable to apply logic that determines the designs to render.
The new design has also removed the 'add and remove involved staff' functionality as it will be added to new edit pages in the application.
The add-involved-staff.cy.js and remove-involved-staff.cy.js have been skipped rather than deleted just so the tests can be referenced when the edit pages and implemented.

* MAP-2319 update db query in line with changes to report_edit table in db

New report owner name and user id will be saved as a concatenated string in the DB.
Access feature flag to control visibility of the new views